# Conflicts:
#	src/main/java/at/qe/skeleton/model/Media.java
#	src/main/java/at/qe/skeleton/services/MediaService.java

+++ src/main/java/at/qe/skeleton/ui/controllers/UserDetailController.java
@@ -1,19 +1,18 @@
 package at.qe.skeleton.ui.controllers;
 
-import java.io.Serializable;
-import java.util.List;
-
-import javax.faces.application.FacesMessage;
-import javax.faces.context.FacesContext;
-
+import at.qe.skeleton.model.Borrowed;
+import at.qe.skeleton.model.User;
+import at.qe.skeleton.model.UserRole;
+import at.qe.skeleton.services.BorrowService;
+import at.qe.skeleton.services.UndoRedoService;
+import at.qe.skeleton.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
-import at.qe.skeleton.model.User;
-import at.qe.skeleton.model.UserRole;
-import at.qe.skeleton.services.UndoRedoService;
-import at.qe.skeleton.services.UserService;
+import java.io.Serializable;
+import java.util.Collection;
+import java.util.List;
 
 /**
  * Controller for the user detail view.
@@ -32,6 +31,14 @@ public class UserDetailController implements Serializable {
 	@Autowired
 	private UndoRedoService undoRedoService;
 
+	@Autowired
+	FMSpamController fms;
+
+	@Autowired
+	BorrowService borrowService;
+
+
+
 	/**
 	 * Attribute to cache the currently displayed user
 	 */
@@ -44,7 +51,7 @@ public class UserDetailController implements Serializable {
 	 * targeted by any further calls of {@link #doReloadUser()},
 	 * {@link #doSaveUser()} and {@link #doDeleteUser()}.
 	 *
-	 * @param user
+	 * @param user the user to set
 	 */
 	public void setUser(final User user) {
 		this.user = user;
@@ -54,7 +61,7 @@ public class UserDetailController implements Serializable {
 	/**
 	 * Returns the currently displayed user.
 	 *
-	 * @return
+	 * @return the set user
 	 */
 	public User getUser() {
 		return this.user;
@@ -82,6 +89,8 @@ public class UserDetailController implements Serializable {
 			// TODO: Exception-Handling
 		}
 		this.doReloadUser();
+		fms.info("A new user with the username: "+ username +" was created!");
+		fms.info("Please reload the page.");
 	}
 
 	/**
@@ -98,30 +107,36 @@ public class UserDetailController implements Serializable {
 		this.undoRedoService.addAction(undoRedoService.createAction(userService.loadUser(this.user.getUsername()), user,
 				UndoRedoService.ActionType.EDIT_USER));
 		this.user = this.userService.saveUser(this.user);
-
-		FacesMessage asGrowl = new FacesMessage(FacesMessage.SEVERITY_INFO, "Changes saved!", "");
-		FacesContext.getCurrentInstance().addMessage("asGrowl", asGrowl);
+		fms.info("Changes saved");
 	}
 
 	/**
 	 * Action to delete the currently displayed user.
 	 */
 	public void doDeleteUser() {
-		try {
-			this.userService.deleteUser(this.user);
-			this.undoRedoService.addAction(undoRedoService.createAction(user, UndoRedoService.ActionType.DELETE_USER));
 
-			FacesMessage asGrowl = new FacesMessage(FacesMessage.SEVERITY_WARN, "User was deleted!", "");
-			FacesContext.getCurrentInstance().addMessage("asGrowl", asGrowl);
-		} catch (UserService.UnauthorizedActionException unauthorizedActionException) {
-			System.out.println(unauthorizedActionException.getMessage());
-			// TODO: Exception-Handling
+		Collection<Borrowed> d1 = borrowService.getAllBorrowsByUser(this.user);
+		if (d1.size() > 0) {
+				fms.warn("This user has still borrowed: " + d1.size() + " article(s) and cannot be deleted yet.");
+		} else {
+				try {
+					this.userService.deleteUser(this.user);
+					this.undoRedoService.addAction(undoRedoService.createAction(user, UndoRedoService.ActionType.DELETE_USER));
+
+				} catch (UserService.UnauthorizedActionException unauthorizedActionException) {
+					System.out.println(unauthorizedActionException.getMessage());
+
+				}
+				this.user = null;
+				fms.info("The user was deleted and all his bookmarks and reserved media has been deleted");
+				fms.info("Please reload the page.");
 		}
-		this.user = null;
+
 	}
 
 	public void changeUserRoles() {
 		userService.changeUserRoles(user, newRolesString);
+		fms.info("Role edit done successfully.");
 	}
 
 	public UserService getUserService() {
